backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
supabase
python-multipart
httpx